mais exercicios

+++ src/aula_java01/ex_1011.java
@@ -17,7 +17,7 @@ public class ex_1011 {
         
         Scanner entrada = new Scanner(System.in);
         
-        raio =entrada.nextInt();
+        raio =entrada.nextDouble();
         volume =(4.0/3)* 3.14159 * (raio*raio *raio);
         
         System.out.printf("VOLUME = %.3f\n", volume);
